🎨⚡️moved ExecuteCommand unit test files
