+++ packages/nodes-base/nodes/ExecuteCommand/test/workflow.json
